groups have been refactored for class MVP, need to write the tests still but MVP for class scope is complete sans unit tests
