+++ DsixRPGcompanionBE/serializers/__init__.py
@@ -3,3 +3,4 @@ from .character import CharacterSerializer
 from .character_skill import CharacterSkillSerializer
 from .skill import SkillSerializer
 from .skill_specialization import SkillSpecializationSerializer
+from .character_group import CharacterGroupSerializer, CharacterGroupCharacterSerializer
