+++ dsix.sql
@@ -16,12 +16,24 @@ DELETE FROM sqlite_sequence WHERE name='DsixRPGcompanionBE_characterskill';
 DELETE FROM DsixRPGcompanionBE_skillspecialization;
 DELETE FROM sqlite_sequence WHERE name='DsixRPGcompanionBE_skillspecialization';
 
+DELETE FROM DsixRPGcompanionBE_usergroup;
+DELETE FROM sqlite_sequence WHERE name='DsixRPGcompanionBE_usergroup';
+
+DELETE FROM DsixRPGcompanionBE_usergroup_characters;
+DELETE FROM sqlite_sequence WHERE name='DsixRPGcompanionBE_usergroup_characters';
+
+DELETE FROM DsixRPGcompanionBE_usergroup_users;
+DELETE FROM sqlite_sequence WHERE name='DsixRPGcompanionBE_usergroup_users';
+
+
 SELECT *
 
 -- a little code returns all tables
 SELECT name 
 FROM sqlite_master 
 WHERE type='table';
+
+DROP TABLE IF EXISTS DsixRPGcompanionBE_user
 
 
 SELECT skill_id, character_id, COUNT(*) AS count
